refactor: 구입 금액 유효성 테스트 코드 분리
